api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.34.0
botocore==1.34.0
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.0
exifread==3.0.0
# Authentication dependencies
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-jose[cryptography]==3.3.0
# XML parsing for heritage API
lxml==4.9.3
